AzureFunctionModule/requirements.txt
azure-functions
azure-storage-blob
imageio
numpy
pydicom
scikit-image